requirements.txt
appium-python-client~=3.1.1
selenium~=4.16.0